requirements.txt
absl-py==2.2.2
alembic==1.16.1
anyio==4.9.0
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
databricks-cli==0.18.0
databricks-sdk==0.54.0
Deprecated==1.2.18
docker==6.1.3
entrypoints==0.4
fastapi==0.100.1
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.8.0
google-cloud-monitoring==2.16.0
google-cloud-storage==2.10.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml-dtypes==0.3.2
mlflow==2.22.0
mlflow-skinny==2.22.0
namex==0.0.9
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opt_einsum==3.4.0
optree==0.15.0
packaging==23.2
pandas==2.0.3
pillow==11.2.1
pluggy==1.6.0
proto-plus==1.26.1
protobuf==4.25.7
pyarrow==15.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==1.10.13
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.4
pywin32==310
PyYAML==6.0.2
querystring-parser==1.2.4
regex==2024.11.6
requests==2.31.0
rich==14.0.0
rsa==4.9.1
scikit-learn==1.3.0
scipy==1.15.3
setuptools==80.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.27.0
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.23.2
waitress==2.1.2
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zipp==3.21.0